ConfigWorkSpace 单元测试
